+++ gnfs/polynomial/__init__.py
