+++ tests/test_linear_algebra.py
@@ -1,4 +1,4 @@
-from gnfs.linear_algebra import solve_matrix
+from gnfs.linalg import solve_matrix
 from gnfs.sieve import Relation
